Block Processing implementation
