+++ src/lib/factories/labwareFactory.ts
@@ -125,6 +125,9 @@ export const fetalWasteFactory = unregisteredLabwareFactory.associations({
 export const dualIndexPlateFactory = unregisteredLabwareFactory.associations({
   labwareType: labwareTypes[LabwareTypeName.DUAL_INDEX_PLATE].build(),
 });
+export const preBarcodedFactory = unregisteredLabwareFactory.associations({
+  labwareType: labwareTypes[LabwareTypeName.PRE_BARCODED_TUBE].build(),
+});
 
 export const labwareFactories: Record<
   LabwareTypeName,
@@ -141,4 +144,5 @@ export const labwareFactories: Record<
   [LabwareTypeName.FOUR_SLOT_SLIDE]: fourSlotSlideFactory,
   [LabwareTypeName.FETAL_WASTE_CONTAINER]: fetalWasteFactory,
   [LabwareTypeName.DUAL_INDEX_PLATE]: dualIndexPlateFactory,
+  [LabwareTypeName.PRE_BARCODED_TUBE]: preBarcodedFactory,
 };
